added json schema
