+++ .gitignore
@@ -1,3 +1,5 @@
 __pycache__/
 venv/
 .env
+alembic.ini
+alembic/
