+++ models.py
@@ -1,9 +1,10 @@
 
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, DateTime,ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime,ForeignKey,Boolean
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import JSONB
 #from sqlalchemy_json import mutable_json_type
+from sqlalchemy.ext.mutable import MutableDict
 Base = declarative_base()
 class Author(Base):
     """
@@ -30,6 +31,7 @@ class Book(Base):
     edition = Column(Integer)
     #year = Column(Integer)
     published_at = Column(DateTime)
+    is_published =Column (Boolean, default=False)
     author=relationship("Author", back_populates="books")
 
 class Users(Base):
@@ -44,6 +46,6 @@ class Users(Base):
     avatar = Column(String)
     phone_number = Column(String)
     #meta_data = Column(mutable_json_type(dbtype=JSONB, nested=True))
-    metadata_info = Column(JSONB)
+    metadata_info =Column(MutableDict.as_mutable(JSONB),nullable =False,server_default="{}")
     authors =relationship("Author",back_populates="user")
